+++ btclib/tx/__init__.py
@@ -7,11 +7,11 @@
 #
 # No part of btclib including this file, may be copied, modified, propagated,
 # or distributed except according to the terms contained in the LICENSE file.
-"""Transactions: OutPoint, TxIn, TxOut, Tx, and join_txs."""
+"""Transactions: OutPoint, TxIn, TxOut, Tx, and join."""
 
 from btclib.tx.out_point import OutPoint
-from btclib.tx.tx import Tx, join_txs
+from btclib.tx.tx import Tx, join
 from btclib.tx.tx_in import TxIn
 from btclib.tx.tx_out import TxOut
 
-__all__ = ["OutPoint", "Tx", "TxIn", "TxOut", "join_txs"]
+__all__ = ["OutPoint", "Tx", "TxIn", "TxOut", "join"]